DHUD for 3.0, 1.5.30000b
Fixed the the issues that 3.0 brought to light. Added basic support for Death Knights.

This is a version of Drathal's HUD that fixes it for use in 3.0.

Nothing is changed from his mod except to fix the the issues that 3.0 brought to light.

Added basic support for Death Knights.

To do: 1. Add Death Knight Runes to the left, like it's done with combo points. 2. Show vehicle health in the pet hud(some vehicles shown here and some are not) 3. Change how pet energy is updated. In this version it's updated on PetEnergyUpdate event instead of every 0.1 sec.